Refactor Ruby plugin
